more qa refinery tweaks
