+++ src/apollo/server/resolvers/locationResolver.js
@@ -15,6 +15,7 @@ import sortByName from './../../../utils/sortByName';
 import setTodaysHours from './../../../utils/setTodaysHours';
 import paginateResults from './../../../utils/paginateResults';
 import filterByTerms from './../../../utils/filterByTerms';
+import addLocationTerms from './../../../utils/addLocationTerms';
 
 // Create a dayjs date object, using default timezone.
 // @see https://github.com/iamkun/dayjs/issues/1227
@@ -22,204 +23,15 @@ const now = dayjs.tz();
 // Format datetime in ISO8601, i.e, 2020-10-27T12:00:00-04:00.
 const today = now.format();
 
-function addLocationTerms(allLocations) {
-  return allLocations.map(location => {
-    // Temp workaround for division & centers that has terms: null
-    if (location.terms === null) {
-      location.terms = [];
-    }
-    // Boroughs.
-    switch (location.locality) {
-      case 'New York':
-        const boroughsManhattan = {
-          uuid: 'filter-borough',
-          name: 'Borough',
-          terms: [
-            {
-              uuid: 'a9405ac3-6d14-4e2b-b8fa-bede59d231b5',
-              name: 'New York'
-            }
-          ]
-        };
-        location.terms.push(boroughsManhattan);
-        break;
-      case 'Bronx':
-        const boroughsBronx = {
-          uuid: 'filter-borough',
-          name: 'Borough',
-          terms: [
-            {
-              uuid: '12060965-2125-4cf9-a724-b3560fdc3af6',
-              name: 'Bronx',
-            }
-          ]
-        };
-        location.terms.push(boroughsBronx);
-        break;
-      case 'Staten Island':
-        const boroughsStatenIsland = {
-          uuid: 'filter-borough',
-          name: 'Borough',
-          terms: [
-            {
-              uuid: 'ecf1cc55-5591-45cc-adf1-6f72aa54f2ce',
-              name: 'Staten Island'
-            }
-          ]
-        };
-        location.terms.push(boroughsStatenIsland);
-        break;
-    }
-
-    // Accessibility
-    switch (location.access) {
-      case 'Fully Accessible':
-        const accessibilityFull = {
-          uuid: 'filter-accessibility',
-          name: 'Accessibility',
-          terms: [
-            {
-              uuid: '4f2f8b1b-dc71-4e66-9657-1c9fba52f3ff',
-              name: 'Fully accessible'
-            },
-          ]
-        };
-        location.terms.push(accessibilityFull);
-        break;
-      case 'Partially Accessible':
-        const accessibilityPartial = {
-          uuid: 'filter-accessibility',
-          name: 'Accessibility',
-          terms: [
-            {
-              uuid: '907db826-860b-4a49-9a4f-e85bf842e193',
-              name: 'Partially accessible'
-            },
-          ]
-        };
-        location.terms.push(accessibilityPartial);
-        break;
-      case 'Not Accessible':
-        const accessibilityNone = {
-          uuid: 'filter-accessibility',
-          name: 'Accessibility',
-          terms: [
-            {
-              uuid: '53c13a59-2aa2-4573-8af4-72fd8df5200a',
-              name: 'Not accessible'
-            },
-          ]
-        };
-        location.terms.push(accessibilityNone);
-        break;
-    }
-  });
-}
-
 const locationResolver = {
   Query: {
     allLocations: async (parent, args, { dataSources }) => {
       const allLocations = await dataSources.refineryApi.getAllLocations();
       let results;
       let totalResultsCount = Object.keys(allLocations).length;
-
-      // Add in hardcoded taxonomy data to each location.
-      // @TODO See if greg can just add this to the refinery output for each location.
-      // if not, move this to a utils function.
-      /*
-      allLocations.map(location => {
-        // Temp workaround for division & centers that has terms: null
-        if (location.terms === null) {
-          location.terms = [];
-        }
-        // Boroughs.
-        switch (location.locality) {
-          case 'New York':
-            const boroughsManhattan = {
-              uuid: 'filter-borough',
-              name: 'Borough',
-              terms: [
-                {
-                  uuid: 'a9405ac3-6d14-4e2b-b8fa-bede59d231b5',
-                  name: 'New York'
-                }
-              ]
-            };
-            location.terms.push(boroughsManhattan);
-            break;
-          case 'Bronx':
-            const boroughsBronx = {
-              uuid: 'filter-borough',
-              name: 'Borough',
-              terms: [
-                {
-                  uuid: '12060965-2125-4cf9-a724-b3560fdc3af6',
-                  name: 'Bronx',
-                }
-              ]
-            };
-            location.terms.push(boroughsBronx);
-            break;
-          case 'Staten Island':
-            const boroughsStatenIsland = {
-              uuid: 'filter-borough',
-              name: 'Borough',
-              terms: [
-                {
-                  uuid: 'ecf1cc55-5591-45cc-adf1-6f72aa54f2ce',
-                  name: 'Staten Island'
-                }
-              ]
-            };
-            location.terms.push(boroughsStatenIsland);
-            break;
-        }
-
-        // Accessibility
-        switch (location.access) {
-          case 'Fully Accessible':
-            const accessibilityFull = {
-              uuid: 'filter-accessibility',
-              name: 'Accessibility',
-              terms: [
-                {
-                  uuid: '4f2f8b1b-dc71-4e66-9657-1c9fba52f3ff',
-                  name: 'Fully accessible'
-                },
-              ]
-            };
-            location.terms.push(accessibilityFull);
-            break;
-          case 'Partially Accessible':
-            const accessibilityPartial = {
-              uuid: 'filter-accessibility',
-              name: 'Accessibility',
-              terms: [
-                {
-                  uuid: '907db826-860b-4a49-9a4f-e85bf842e193',
-                  name: 'Partially accessible'
-                },
-              ]
-            };
-            location.terms.push(accessibilityPartial);
-            break;
-          case 'Not Accessible':
-            const accessibilityNone = {
-              uuid: 'filter-accessibility',
-              name: 'Accessibility',
-              terms: [
-                {
-                  uuid: '53c13a59-2aa2-4573-8af4-72fd8df5200a',
-                  name: 'Not accessible'
-                },
-              ]
-            };
-            location.terms.push(accessibilityNone);
-            break;
-        }
-      });
-      */
-      allLocations = addLocationTerms(allLocations);
+      
+      // Temp workaround to get location terms data.
+      addLocationTerms(allLocations);
       
       // Filter by openNow
       if (
@@ -287,6 +99,8 @@ const locationResolver = {
   Location: {
     id: location => location.slug.replace('/', '-'),
     name: location => location.name,
+    contentType: location => location.type,
+    slug: location => location.slug,
     status: location => location.slug,
     address_line1: location => location.street_address,
     address_line2: location => location.street_address,
