+++ src/apollo/server/type-defs.js
@@ -4,6 +4,8 @@ export const typeDefs = gql`
   type Location {
     id: ID!
     name: String!
+    contentType: String!
+    slug: String!
     status: String!
     address_line1: String
     address_line2: String
